requirements.txt
Flask~=1.1.2
flask-restx~=0.2.0
flask-redis~=0.4.0
Flask-PyMongo~=2.3.0
flask-mongoengine~=1.0.0
mongoengine~=0.22.1
flask-marshmallow~=0.14.0
Flask-JWT-Extended~=3.25.0
Flask-Cors~=3.0.9
Flask-Bcrypt~=0.7.1
Flask-BasicAuth~=0.2.0
websocket-client~=0.57.0
typed-ast~=1.4.2
timeout-decorator~=0.5.0
temporary~=3.0.0
rq-dashboard~=0.6.1
pyOpenSSL~=20.0.1
gunicorn~=20.0.4
gevent~=20.9.0
blinker~=1.4
billiard~=3.6.3.0
executor~=23.2
kombu~=5.0.2
numpy~=1.19.4
matplotlib~=3.3.3
networkx~=2.5

# pi==0.1.2
# pip==20.3.3
# py4j==0.10.9.1
# paramiko==2.7.2
# monotonic==1.5
# email-validator==1.1.2
# importlib-metadata==3.1.0

